pomodoro tidy - renamed helper function and removed prev Pomodoro versions
