+++ src/sub-pages/Pomodoro.js
@@ -1,7 +1,4 @@
-/* import PomodoroTimer from '../components/PomodoroTimerAbsolute'; */    
-/* import PomodoroTimer from '../components/PomodoroTimer'; */
-/* import PomodoroTimer from '../components/PomodoroTimerNew'; */
-import PomodoroTimer from '../components/PomodoroTimer1';
+import PomodoroTimer from '../components/PomodoroTimer';
 
 import imgBg from '../assets/LandingPage.png';
